Improved files removing behavior is case of file locked.

+++ fr.ocelet.datafacer/src/fr/ocelet/datafacer/ocltypes/Shapefile.java
@@ -103,30 +103,31 @@ public abstract class Shapefile extends GtDatafacer
 	 */
 	public void setFileName(String shpFileName) {
 		sourceFile = new File(FileUtils.applyOutput(shpFileName));
-setDatastore();
-//		try {
-//			// Open a datastore that uses our own geometry factory
-//			datastore = new OcltShapefileDataStore(sourceFile.toURI().toURL());
-//			String typeName = datastore.getTypeNames()[0];
-//			if (sourceFile.exists()) {
-//			FeatureSource source = datastore.getFeatureSource(typeName);
-//				bounds[0] = source.getBounds().getMinX();
-//				bounds[1] = source.getBounds().getMinY();
-//				bounds[2] = source.getBounds().getMaxX();
-//				bounds[3] = source.getBounds().getMaxY();
-//			}
-//		} catch (IOException e) {
-//			System.out.println(ERR_HEADER + "Failed to open the shapefile " + sourceFile);
-//		}
+		setDatastore();
+		// try {
+		// // Open a datastore that uses our own geometry factory
+		// datastore = new OcltShapefileDataStore(sourceFile.toURI().toURL());
+		// String typeName = datastore.getTypeNames()[0];
+		// if (sourceFile.exists()) {
+		// FeatureSource source = datastore.getFeatureSource(typeName);
+		// bounds[0] = source.getBounds().getMinX();
+		// bounds[1] = source.getBounds().getMinY();
+		// bounds[2] = source.getBounds().getMaxX();
+		// bounds[3] = source.getBounds().getMaxY();
+		// }
+		// } catch (IOException e) {
+		// System.out.println(ERR_HEADER + "Failed to open the shapefile " +
+		// sourceFile);
+		// }
 	}
-	
+
 	protected void setDatastore() {
 		try {
 			// Open a datastore that uses our own geometry factory
 			datastore = new OcltShapefileDataStore(sourceFile.toURI().toURL());
 			String typeName = datastore.getTypeNames()[0];
 			if (sourceFile.exists()) {
-			FeatureSource source = datastore.getFeatureSource(typeName);
+				FeatureSource source = datastore.getFeatureSource(typeName);
 				bounds[0] = source.getBounds().getMinX();
 				bounds[1] = source.getBounds().getMinY();
 				bounds[2] = source.getBounds().getMaxX();
@@ -290,11 +291,17 @@ setDatastore();
 	@Override
 	public void remove() {
 		String rootName = sourceFile.getPath().substring(0, sourceFile.getPath().indexOf(".shp"));
-		Miscutils.removeFile(rootName + ".prj");
-		Miscutils.removeFile(rootName + ".shx");
-		Miscutils.removeFile(rootName + ".shp");
-		Miscutils.removeFile(rootName + ".dbf");
-		setDatastore(); // needed to be able to write again content after removing the files.
+		if ((Miscutils.isLocked(rootName + ".shx")) || (Miscutils.isLocked(rootName + ".shp"))
+				|| (Miscutils.isLocked(rootName + ".dbf")) || (Miscutils.isLocked(rootName + ".prj")))
+			System.out.println("Warning: the shapefile " + sourceFile + " could not be removed. It may be in use by another program.");
+		else {
+			Miscutils.removeFile(rootName + ".prj");
+			Miscutils.removeFile(rootName + ".shx");
+			Miscutils.removeFile(rootName + ".shp");
+			Miscutils.removeFile(rootName + ".dbf");
+			setDatastore(); // needed to be able to write again content after
+							// removing the files.
+		}
 	}
 
 	@Override
